+++ src/components/BackOfficeComponent.tsx
@@ -1,9 +1,9 @@
-import itImage from '../../../images/ITA_Logo.png'
-import LogoOut from '../../../images/Sign-out.png'
+import itImage from '../images/ITA_Logo.png'
+import LogoOut from '../images/Sign-out.png'
 
 import { useEffect, useState } from "react";
-import FAQs from "../../landingPage/components/faqs/FAQsComponent";
-import ProjectsComponent from "../../landingPage/components/apps/ProjectsComponent";
+import FAQs from "./faqs/FAQsComponent";
+import ProjectsComponent from "./apps/ProjectsComponent";
 
 function ViewBackOffice({setIsLogged,dispatch}:{setIsLogged:any,dispatch:any}) {
 
@@ -13,7 +13,7 @@ function ViewBackOffice({setIsLogged,dispatch}:{setIsLogged:any,dispatch:any}) {
     });
     useEffect(() => {
         setState({...state, faqs: false});
-    }, []);
+    }, [ ]);
