+++ src/components/SloganComponent.tsx
@@ -1,13 +1,13 @@
-import angularLogo from '../../../../img/angular.png';
-import phpLogo from '../../../../img/php.png';
-import javaLogo from '../../../../img/java.png';
-import pythonLogo from '../../../../img/python.png';
-import reactLogo from '../../../../img/react.png';
-import scrumLogo from '../../../../img/scrum.png';
-import nodejsLogo from '../../../../img/nodejs.png';
-import gitLogo from '../../../../img/git.png';
-import itacademyLogo from '../../../../img/itacademylogo.png';
-import HeaderComponent from '../header/HeaderComponent'
+import angularLogo from '../img/angular.png';
+import phpLogo from '../img/php.png';
+import javaLogo from '../img/java.png';
+import pythonLogo from '../img/python.png';
+import reactLogo from '../img/react.png';
+import scrumLogo from '../img/scrum.png';
+import nodejsLogo from '../img/nodejs.png';
+import gitLogo from '../img/git.png';
+import itacademyLogo from '../img/itacademylogo.png';
+import HeaderComponent from './HeaderComponent'
 
 
 const SloganComponent = () => (
