+++ src/components/apps/ProjectsComponent.tsx
@@ -1,48 +1,38 @@
-import curvedArrow from "../../../../img/curved-arrow.png"
+import curvedArrow from "../../img/curved-arrow.png"
 import { FaRegCircle, FaArrowRight } from "react-icons/fa";
 import { IconContext } from "react-icons";
 import { useEffect,useState } from "react";
-import { apiCallApps, apiCallAppsInfo,postApiApps } from "../../../../store/reducers/appsCall/appsCallApiFunctionality";
+import { apiCallApps, apiCallAppsInfo,postApiApps } from "../../store/reducers/appsCall/appsCallApiFunctionality";
 import { useDispatch, useSelector } from "react-redux";
-import { RootState } from "../../../../store/store";
-import Pencil from '../../../../img/vector-10.png'
-import ModalApps from "../../../backOffice/components/appsAdminView/modalApps";
-import ModalsAddApps from '../../../backOffice/components/appsAdminView/ModalsAddApps'
+import { RootState } from "../../store/store";
+import { createToken, ApiStateApps} from "../../interfaces/interfaces";
+import Pencil from '../../img/vector-10.png'
+import ModalApps from "./appsAdminView/modalApps";
+import ModalsAddApps from './appsAdminView/ModalsAddApps'
 
 declare global {
   interface Window {
     my_modal_1: {
       showModal: () => void;
     };
-  }
-}
-
-declare global {
-  interface Window {
     my_modal_2: {
       showModal: () => void;
     };
   }
 }
-export const ProjectsComponent = () => {
 
-  interface createToken {
-    acces_token: string;  
-  }
+export const ProjectsComponent = () => {
 
   const { acces_token }: createToken = useSelector(
     (state: RootState) => state.apiPostRegister
   );
 
-  const handleSendApiInfo= (id:any)=>{ 
+  const handleSendApiInfo= (id:number)=>{ 
     apiCallAppsInfo(dispatch,id,acces_token)
   };
-
-  interface ApiPostRegisterState {
-    apps: any[]; 
-  }
   
-  const { apps }: ApiPostRegisterState = useSelector(
+  
+  const { apps }: ApiStateApps = useSelector(
     (state: RootState) => state.appsCallApiFunctionality
   );
