+++ src/pages/BackOffice.tsx
@@ -1,6 +1,6 @@
 import { useDispatch } from "react-redux";
-import { setIsLogged } from "../../store/reducers/apiCall/apiPostRegisterLogin";
-import ViewBackOffice from "./components/ViewBackOffice";
+import { setIsLogged } from "../store/reducers/apiCall/apiPostRegisterLogin";
+import ViewBackOffice from "../components/BackOfficeComponent";
 function BackOffice() {
   const dispatch = useDispatch()
   return (
