+++ src/router/routes.tsx
@@ -1,6 +1,6 @@
 import { Route, Routes} from "react-router-dom"
-import LandingPage from "../pages/landingPage/LandingPage"
-import BackOffice from "../pages/backOffice/BackOffice"
+import LandingPage from "../pages/LandingPage"
+import BackOffice from "../pages/BackOffice"
 import ProtectedRoute from "./protectedRoutes"
 
 function Router() {
