+++ src/store/reducers/appsCall/appsCallApiFunctionality.tsx
@@ -1,28 +1,13 @@
 import { createSlice } from '@reduxjs/toolkit'
-import axios from 'axios'
-
-
-
-interface faqsData {
-apps:any[],
-appsInfo:[]
-gitURL:string,
-repoURL:string,
-stateProject:string
-loadingApps:boolean
-}
-
-
-const initialState: faqsData = {
+import axios from 'axios' 
+const initialState = {
   apps:[],
   appsInfo:[],
   gitURL:'',
   repoURL:'',
   stateProject:'',
-  loadingApps:false
-  
-}
-
+  loadingApps:false 
+} 
 export const apiSlice = createSlice({
   name: 'appsCallApiFunctionality',
   initialState,
@@ -52,7 +37,7 @@ export const { setApps,setAppsInfo,setLoadingApps } = apiSlice.actions;
     
   }
   //llamada api Apps ID//
-  export const apiCallAppsInfo =async(dispatch:any,id:number,acces_token:String)=>{
+  export const apiCallAppsInfo =async(dispatch:any,id:number,acces_token:string)=>{
     try {
       dispatch(setLoadingApps(true));
 
@@ -72,7 +57,7 @@ export const { setApps,setAppsInfo,setLoadingApps } = apiSlice.actions;
     
   }
   //Post //
- export const postApiApps =async(faqContent:any,acces_token:String, dispatch:any)=>{
+ export const postApiApps =async(faqContent:any,acces_token:string, dispatch:any)=>{
     try{
         
      await axios.post('http://87.106.229.119/api/apps',faqContent,{
@@ -88,7 +73,7 @@ export const { setApps,setAppsInfo,setLoadingApps } = apiSlice.actions;
   };
   //delete //
 
-  export const deleteApiApps = async (appsID:number, acces_token:String,dispatch:any) => {
+  export const deleteApiApps = async (appsID:number, acces_token:string,dispatch:any) => {
     try {
       const url = `http://87.106.229.119/api/apps/${appsID}`;
       await axios.delete(url,{
@@ -106,7 +91,7 @@ export const { setApps,setAppsInfo,setLoadingApps } = apiSlice.actions;
   };
 //put//
 
-export const putApiApps =async(faqContent:any,acces_token:String, dispatch:any, id:number)=>{
+export const putApiApps =async(faqContent:any,acces_token:string, dispatch:any, id:number)=>{
   try{
      
    await axios.put('http://87.106.229.119/api/apps/'+id,faqContent,{
